CASMTRIAGE-3591 Reduce kdump script
The overlayFS post-script and fstab were causing problems, this change
refactors the kdump script to use straight partitions.

There are some other changes in here that remove some layers that aren't
needed, and there are updated hacks.

The hack at present regenerates the initrd and fixes the fstab and the
mountpoints for kdump.