MTL-1842 Fix output and fstab formatting
This amends a newline to the fail log line, ensuring the 'See logfile
at' line isn't awkwardly appended to the end of the status block.

This also fixes the invalid metal-iptables prints, which were defined
using more than 20 characters without modifying the printf line to
match. This change conforms metal-iptables printf to 20 chars so it
nicely sits with the other printf calls.

Lastly this replaces the static hardcoding of the bootloader filesystem
label with a dynamic resolution.