MTL-1841 Native kdump (#18)
Remove `create-kdump-artifacts.sh` and `kdump-cray.service`, neither are
needed with the native kdump approach.

Remove `05-metal.conf` and extra dracut call. The `05-metal.conf` is in
the way when kdump runs and the return on its investment is diminishing
as time goes on. We do not want to include the `fstab.metal` file anymore
when an image is created, in most contexts when we recreate an image we
just want it to be the bare minimum without concessions to the node it
was built on.

Include `metalfs.service` in all NCN image types in order to mount
`/var/crash` on storage nodes. Originally `metalfs.service` was missing
from storage nodes because they did not have an `/etc/fstab.metal` file,
but they have had one for a while now. Technically storage nodes were
never mounting `/metal/recovery` automatically since `metalfs.service`
was missing. To make `metalfs.service` more friendly, it now has an
`ExecCondition` that starts the service _only_ when `/etc/fstab.metal`
exists.

This also removes the extra `dracut --force` call, I am not sure why it
existed but it seems like extra. Some ancilliary changes for
reorganizing the files in `files/` is included because of the new
`00-csm.conf` file.

Lastly this moves contents from `dracut` calls and `dracut-lib.sh` into
a `.conf` file.