MTL-1876 Install initrd and kernel
Install the initrd and kernel from the booted squashFS instead of
relying on dracut to fetch it.

This is needed for pre-signed URLs to work, because we won't be able to
fetch a pre-signed URL for the initrd and kernel during dracut. The
pre-signed URL for the initrd and kernel are only available to iPXE.

Update the `create-ims-initrd.sh` script to populate `/squashfs` so the
bootloader code can grab up-to-date artifacts from IMS builds.