MTL-1476 Move `create-kis-artifacts.sh` to `metal/`
MTL-1476 is adding a new script in `node-images` that requires
`create-kis-artifacts.sh`, the new script is for metal only. The
`create-kis-artifacts.sh` script technically is also metal only, so to
prevent a confusing connection of a metal-only script requiring another
metal-only script that lives in `common/` this is moving
`create-kis-artifacts.sh` and its cleanup script to `metal/`.

This also updates the copyrights.