VSHA-536 Do not rebuild kdump's initrd in chroot
When running `create-kis-artifact.sh` in a chroot, a common thing to do
when developing on a Cray non-compute node, do not rebuild the kdump
initrd.

The chrooted environment is missing a few mount points that causes the
kdump initrd to build in non-hostonly mode, and in doing so it
disregards our request to use the fstab. This results in a kdump initrd
that has an unreadable (blank) `/etc/fstab`, preventing a dump from
working:

```bash
[    0.067767][    T1] [Firmware Bug]: the BIOS has corrupted hw-PMU resources (MSR 38d is b0)
[    4.349296][    T1] mce: Unable to init MCE device (rc: -5)
Cannot save dump!

  Can't read fstab.

Something failed. You can try to debug it here.
Type 'reboot -f' to reboot the system or 'exit' to
resume the boot process.
sh-4.4#
```