Use an upperbound for ceph
Artifactory had ceph RPMs for version 16 hiding in the csm-rpms RPM
repository, and since that repository had a higher priority than the
filesystems-ceph one we had to use the `>` or `>=` operator to force
selection of `ceph-17`. However this had a pitfall, now that ceph-17.2.6
is available its packages were being chosen over our desired 17.2.5
RPMs.

We've removed the ceph RPMs from csm-rpms, which were uploaded by Pete a
long time ago as an attempted workaround for a CSM tarball. In doing so,
using `ceph<17.2.6` now properly resolves 17.2.5 (instead of the old
16.2.9 RPMs).