CASMPET-6542 New `loftsman` and `helm-bash-completion`
The new `1.2.0-2` `loftsman` no longer bundles `helm`, which allows us
to install the `helm` package (thus making it known to the running
Zypper instance). This also enables us to install
`helm-bash-completion`, which requires the `helm` package.

Previously `helm-bash-completion` could not install as it required the
`helm `package, but the `helm` package conflicted with `loftsman` due to
the bundling.