MTL-2032 cloud-init packages module
Update `setup-nexus.sh` with the ability to upload CSM tarball
reposiroties, or any directory, into nexus.

Update apache2 configuration to proxy https://packages.nmn to the
pit-nexus for the pre-install-toolkit. This will allow the repository
URLs to be the same when served by the PIT or K8s.

Ensure cloud-init configs from NCN Common apply to all contexts now that
vshastav2 is running metal artifacts.