Adds `pulumi` with spice support
Installs `pulumi`, an infrastructure as code application for manageing
VM deployments. This application requires `python310` and `go1.20`;
the former is already installed in our image, the latter is now
installed alongside `pulumi`.

Also includes `qemu-ui-spice-core`, a needed dependency for setting up
monitors on the VMs. This dependency fixes the `spice support is
disabled` error.