MTL-2032 Fix group repo names
The `exists` variable was checking the member name instead of the group
repo name, this was leading the code to believe the repo group already
existed and it was doing a `PATCH` which doesn't work.

This fixes the code to properly use `repo_group_name where necessary,
and successfully creates the repo group.