MTL-2189 Break out packages; mitigate CVEs in hypervisor
The hypervisor and new VM images are pulling in node-images-base, which
is installing several items that we do not want in those new images.

In an attempt to appease the new and current collection of images, this
change breaks out the CSM packages and a curated list of user-space
packages into their own, respective roles.

The CSM packages are now opt-in, as well as the user-space packages.
This will significantly slim down the CVE footprint of our new images,
while allowing us to have a common ground for all images.
node-images-base now only contains a curated, bare-minimum set of
packages. Each package in node-images-base also has its own rationale
memo.