MTL-2192 Remove/overwrite Duplicate FSLabel
The ISO has a duplicate label which is causing problems when booting
from a USB when using systemd 249.16-150400.8.28.3 and newer. SUSE has
changed the behavior of the 60-persistent-storage udev rules, which no
longer plays nice with the implicitly set duplicate label.

Writing a different label to the first few sectors of the ISO prevents
the duplicate label from appearing, and allows the correct partition to
be selected.