MTL-2209 Change default socket buffer size from 256M to 16M (#316)
There is a bug in "weave" where it does not drain packets from
the UDP sockets that it creates for each peer, thus tying up
system memory.  And because this is not a STREAM socket, the
actual memory consumed is more than the 256M of user data, it
is roughly 1.5 times, or about 384MB per UDP socket.  This is
an N^2 issue, so for a 40 node system management cluster there
will be over 500GB of memory tied up in these sockets across
the system.

The mitigation is to reduce the default socket buffer size from
256MB to 16MB, which is still more than enough.  The maximum
socket buffer size is left at 256MB, so individual applications
can still use larger sizes if they really need to.

In general, for a TCP connection, to maintain full speed
throughput and be able to sustain one packet loss per RTT,
you need twice the delay*bandwidth for the socket buffer.  To
sustain 2 losses/RTT, three times the delay*bandwith.

Our systems internally have sub-millisecond RTTs, so even at
10Gbit/sec with 2 packet losses per RTT, you'd still only need
less than a 4MB buffer to sustain full throughput.