The k8s-vm had out-of-date cloud-init Ansible
The Ansible for the management-vm and kubernetes-vm regarding cloud-init
files should be the same.

The node-image docs were expecting to find `00_network.cfg` for the
kubernetes-vm in a similar directory to the management-vm.

The kubernetes-vm `cloud.cfg` was a duplicate that wasn't cleaned up
when the `cloud_init` role came to be.

(cherry picked from commit a4bdfe3ef9b095d044a9fb1667b64c16a19a5631)