MTL-2253 Add SP5 RPMs with aarch64
The previously used PTF 1213185 RPMs were for SLE-15-SP4, despite these
working in testing it would be more appropriate to replace them with SP5
variaents. The new SP5 RPMs have a different PTF number of 1214447.

The new repo URL also contains aarch64 RPMs, which were missed in the
recent PTF addition. This will unblock aarch64 builds.

Lastly, I reorganized Artifactory to use a different directory structure
for the PTF packages. The URLs for the previous PTF and new PTF are
updated to reflect this.