CASMTRIAGE-5988 Fix quotes
When `ISO_FSLABEL` is not defined, a default value of `'CRAYLIVE'` is
set with single-quotes. The single-quotes create a malformed disk label:

```
/dev/sdd1: BLOCK_SIZE="2048" UUID="2023-08-31-19-27-57-00" LABEL="'CRAYLIVE'" TYPE="iso9660" PARTLABEL="ISO9660" PARTUUID="6aaecd22-29dd-4800-b511-b53fe40d5b8e"
```

Specifically `LABEL="'CRAYLIVE'"`, this is difficult to match and our
bootloaders do not expect the quotes to exist in the FSLabel. This
causes a failure to boot:

```
Warning: /dev/disk/by-label/CRAYLIVE does not exist
```