CASMTRIAGE-5988 Empty `root=live:LABEL=` argument
`ISO_FSLABEL` is not exported by default, and as a result it is not
rendored during the `envsubst` command when generating `grub.cfg`.

Packer builds that set `ISO_FSLABEL` in their environment will
workaround this behavior because Packer exports its environment
variables into the script. The pre-install-toolkit build was not setting
`ISO_FSLABEL` in the scripts provisoner environment, so it was not being
exported.

Adding an `export` into the script will ensure the variable is always
exported and rendored in `grub.cfg`.

(cherry picked from commit 555ca6eb5bf8075f1855a27d806b083af7446fcc)