MTL-2000 Enable zypper cloud-init modules
Enables repository setup and package installation from cloud-init, using
the zypper and packages cloud-init modules.

This deprecates and disables the package install hack
(`install_csm_rpms`).

This change depends on the following cloud-init user-data existing:

```json
{
  "repos": [
    "cani",
    "canu",
    "cray-cmstools-crayctldeploy",
    "csm-testing",
    "goss-servers",
    "iuf-cli",
    "libcsm",
    "platform-utils"
  ],
  "zypper": {
    "repos": [
      {
        "autorefresh": 1,
        "baseurl": "https://packages.local/repository/csm-noos?ssl_verify=no",
        "enabled": 1,
        "gpgcheck": 0,
        "id": "csm-noos",
        "name": "csm-noos"
      },
      {
        "autorefresh": 1,
        "baseurl": "https://packages.local/repository/csm-sle-${releasever_major}sp${releasever_minor}?ssl_verify=no",
        "enabled": 1,
        "gpgcheck": 0,
        "id": "csm-sle",
        "name": "csm-sle-${releasever_major}sp${releasever_minor}"
      }
    ]
  }
}
```
The above snippet was confirmed on a csm-1.5.0-beta.64 upgrade, and was
confirmed on a CSM 1.5 fresh install at an earlier date.

The Zypper repo information in the snippet was used manually and
resulted in successfully adding the repository.