CRAYSAT-1774: Update version of cray-sat-podman and remove sat-rpms repo
The `cray-sat-podman` RPM has been moved from the sat-rpms repository in
Artifactory to the csm-rpms repository. Its version has also been
incremented from 2.0.0-1 to 2.1.0-1. Update the version of this package
included in the Kubernetes NCN images, and remove the sat-rpms
repository, which is no longer used for any RPMs included here.