Configure DNS on the local hypervisor
The DNS options are not fulfilled by libvirt's DHCP configuration.

Instead of cramming more configuration into libvirt, and since we don't care about DNS until after this role runs, the network_manager role can configure DNS.

This configures DNS for the local hypervisor node, setting `ipv4.dns` and `ipv4.dns-search` to match the DHCP lease for the management VM and the search domain for the metal network.

The search domain is hardcoded, which is bad. It could be fixed by moving the value out of this role and the dnsmasq role into some global variable.