MTL-2414 initrd and kdump initrd additions
Customer has requested that both the kdump and normal initrds for NCNs include more debugging binaries as well as the squashFS kernel module.

While the squashfs kernel module was present in our main initrd, it was not present in the kdump initrd.

This change adds the requested items to our dracut configuration, forcing both `dracut` and `mkdumprd` to include them.

New items:
- `squashfs.ko` (kdump initrd)
- `dmidecode`
- `gzip`
- `lsof`
- `lspci`
- `lsscsi`
- `md5sum`
- `netcat`
- `tar`
- `vim`
- `xz`

NOTE: In order to include `netcat`, the `netcat-openbsd` was needed. This package was added to all node-images via `vars/packages/suse.yml`.